requirements.txt
fusepy
click